tests(cache) improve condition for regression test for #3277
This changes the regression test to make it not dependent on an
absolute memory threshold value, but instead on the general
observed behavior of the GC: the test fails if used memory
increases (by any amount) on every iteration of the outer loop.

In local tests, this consistently fails without the leak fix
and passes with the leak fix.